refactor json tests into multiple files
